+++ commons-json/src/test/java/com/hb0730/commons/json/jackson/JacksonImplTest.java
@@ -1,10 +1,18 @@
 package com.hb0730.commons.json.jackson;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
+import com.fasterxml.jackson.databind.type.TypeFactory;
 import com.hb0730.commons.json.exceptions.JsonException;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Assert;
 import org.junit.Before;
@@ -12,6 +20,7 @@ import org.junit.Test;
 
 import java.io.IOException;
 import java.io.Serializable;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -55,7 +64,7 @@ public class JacksonImplTest {
         log.info(test1.toString());
         Test1 test11 = jacksons.jsonToObject(null, Test1.class, mapper);
         Assert.assertNotNull(test11);
-        Test1 test111 = jacksons.jsonToObject(null, null, mapper);
+        Test1 test111 = jacksons.jsonToObject(null, (Class<? extends Test1>) null, mapper);
         Assert.assertNotNull(test111);
     }
 
@@ -150,6 +159,26 @@ public class JacksonImplTest {
         log.info(map.toString());
     }
 
+    //------------------------泛型擦除问题
+    @Test
+    public void test() throws JsonProcessingException {
+        Test2 test1 = new Test2();
+        test1.setId("1");
+        test1.setName("测试1");
+        Test2 test2 = new Test2();
+        test2.setId("2");
+        test2.setName("测试2");
+        Page<Test2> page1 = new Page<>();
+        page1.setData(Arrays.asList(test1, test2));
+        page1.setName("test");
+
+        String jsonStr = jacksons.objectToJson(page1);
+        Page<Test2> page2 = jacksons.jsonToObject(jsonStr, Page.class);
+
+        JavaType javaType = TypeFactory.defaultInstance().constructParametricType(Page.class, Test2.class);
+        Page<Test2> page3 = jacksons.jsonObject(jsonStr, javaType);
+    }
+
     @Data
     @Builder
     @ToString
@@ -159,5 +188,17 @@ public class JacksonImplTest {
     static class Test1 implements Serializable {
         private String id;
         private String name;
+    }
+
+    @Data
+    public static class Test2 implements Serializable {
+        private String id;
+        private String name;
+    }
+
+    @Data
+    public static class Page<T> implements Serializable {
+        private String name;
+        private List<T> data;
     }
 }